pyproject.toml
[project]
name = "ragicane"
version = "0.1.0"
description = "Read Tropical Storm reports from NHC and talk to them"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.2",
    "aioresponses>=0.7.8",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "nltk>=3.9.1",
    "pdfminer-six>=20250506",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
    "spacy>=3.8.7",
]
[tool.uv]
package = true
[project.scripts]
ragicane = "ragicane.cli:main"
[tool.setuptools.packages.find]
where = ["."]
exclude = ['sample_reports']
[tool.flake8]
ignore = ["E501"]
exclude = [".venv", "build", "dist"]
# max-line-length = 120
